Implement user and user_manager
This is the code for the foundation of the user and user manager class. It is intended to use
the UserManager class to create a user or a superuser. This is accessible in the `objects` field
in the `User` model (retrievable with `from django.contrib.auth import get_user_model`.