Attempting to enforce styling with EditorConfig
EditorConfig is a VSCode extension that enforces / updates files on save to match a specified ruleset
that can be found in `.editorconfig` in the root of the project.

How to Use:

You will need to install the EditorConfig extension in VSCode. Additionally, you may have to right click
on the work space and select `Generate new editorconfig` for the extension to work properly.