added the WIP Sprint Report 2
I meant to add this report at the start of the sprint to the develop branch, but I waited because I thought PR #34 was going to be merged soon and potentially pushed to main. Since that didn't end up happening at all, I am now pushing my updates so anyone else can add their contributions before the Sprint 2 ends on 4/10.