Restructured the computations package.
The restructured package now separates the configurations into a separate config.json file
and the class model. A list of models can be found in computations/__init__.py in the `COMPUTE_MODELS` global.
Additionally, in the compuations.config.Config class, you can pass in the object of the compute model to any of the
`Config` functions to acquire data in the `config.json` about it