Updating Parameter to be a function, fixed bugs Updates issue #69. - changed Parameter from class to function with hooks (reasoning discussed in 6f0bcc922978f32f851cfb3d41fb6d8416417f94) - fixed bug where parameters were being stored as empty rather than their default values - fixed bug where parameter would not be saved if text was selected all at once and deleted
Now, if the user has entered in a non-default value, the text will be in the parameter field.
If it's still the default value, the text will be empty so as to show the placeholder value (aka the default value).
All while also still saving the correct values in localStorage.